--- doc/GroupingHandler.js.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <title>JSDoc: Source: GroupingHandler.js</title>
-
-    <script src="scripts/prettify/prettify.js"> </script>
-    <script src="scripts/prettify/lang-css.js"> </script>
-    <!--[if lt IE 9]>
-      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
-    <![endif]-->
-    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
-    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
-</head>
-
-<body>
-
-<div id="main">
-
-    <h1 class="page-title">Source: GroupingHandler.js</h1>
-
-    
-
-
-
-    
-    <section>
-        <article>
-            <pre class="prettyprint source linenums"><code>/** Handles grouping of neumes.
- * @constructor
- */
-function GroupingHandler () {
-
-    //TODO: CHANGE NAVABAR-LINK TO PROPER ICON//
-    function triggerGroupSelection () {
-        $("#moreEdit").removeClass("is-invisible");
-        $("#moreEdit").append(
-            "&lt;label>Select Grouping:&amp;nbsp;&lt;/label>" +
-            "&lt;div id='grouping_dropdown' class='dropdown'>" +
-            "&lt;div class='dropdown-trigger'>" +
-            "&lt;button id='select-options' class='button navbar-link' aria-haspopup='true' aria-controls='dropdown-menu'>" +
-            "&lt;span>Groupings&lt;/span>&lt;span class='icon is-small'>" +
-            "&lt;i class=''>&lt;/i>&lt;/span>&lt;/button>&lt;/div>" +
-            "&lt;div class='dropdown-menu' id='dropdown-menu' role='menu'>" +
-            "&lt;div class='dropdown-content'>" +
-            "&lt;a id='Torculus' class='dropdown-item'>Torculus&lt;/a>" +
-            "&lt;a id='' class='dropdown-item'>Etc...&lt;/a>&lt;/div>&lt;/div>&lt;/div>" +
-            "&lt;div>&lt;p class='control'>" +
-            "&lt;button class='button' id='ungroup'>Ungroup&lt;/button>&lt;/p>&lt;/div>"
-
-        );
-        initGroupingListeners();
-    };
-
-    function endGroupingSelection () {
-        $("#moreEdit").empty();
-        $("#moreEdit").addClass("is-invisible");
-    }
-
-    function initGroupingListeners(){
-        $("#grouping_dropdown").on("click", function() {
-            $(this).toggleClass("is-active");
-        })
-
-    }
-    
-    GroupingHandler.prototype.constructor = GroupingHandler;
-    GroupingHandler.prototype.triggerGroupSelection = triggerGroupSelection;
-    GroupingHandler.prototype.endGroupingSelection = endGroupingSelection;
-}
-export {GroupingHandler as default};
-</code></pre>
-        </article>
-    </section>
-
-
-
-
-</div>
-
-<nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Select.html">Select</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Select.ClickSelect.html">ClickSelect</a></li><li><a href="module-Select.DragSelect.html">DragSelect</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li></ul>
-</nav>
-
-<br class="clear">
-
-<footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:46:39 GMT-0400 (EDT)
-</footer>
-
-<script> prettyPrint(); </script>
-<script src="scripts/linenumber.js"> </script>
-</body>
-</html>
