+++ doc/module-Grouping.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>JSDoc: Class: EditMode</title>
+    <title>JSDoc: Module: Grouping</title>
 
     <script src="scripts/prettify/prettify.js"> </script>
     <script src="scripts/prettify/lang-css.js"> </script>
@@ -17,7 +17,7 @@
 
 <div id="main">
 
-    <h1 class="page-title">Class: EditMode</h1>
+    <h1 class="page-title">Module: Grouping</h1>
 
     
 
@@ -28,7 +28,6 @@
 
 <header>
     
-        <h2><span class="attribs"><span class="type-signature"></span></span>EditMode<span class="signature">(neonView, meiFile, zoomHandler)</span><span class="type-signature"></span></h2>
         
     
 </header>
@@ -38,184 +37,7 @@
     
         
 
-    
-
-    
-    <h4 class="name" id="EditMode"><span class="type-signature"></span>new EditMode<span class="signature">(neonView, meiFile, zoomHandler)</span><span class="type-signature"></span></h4>
-    
-
-    
-
-
-
-<div class="description">
-    Creates user interface for editing and creates necessary tools.
-</div>
-
-
-
-
-
-
-
-
-
-    <h5>Parameters:</h5>
-    
-
-<table class="params">
-    <thead>
-    <tr>
         
-        <th>Name</th>
-        
-
-        <th>Type</th>
-
-        
-
-        
-
-        <th class="last">Description</th>
-    </tr>
-    </thead>
-
-    <tbody>
-    
-
-        <tr>
-            
-                <td class="name"><code>neonView</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type"><a href="NeonView.html">NeonView</a></span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The NeonView parent object.</td>
-        </tr>
-
-    
-
-        <tr>
-            
-                <td class="name"><code>meiFile</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type">string</span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The path to the MEi file.</td>
-        </tr>
-
-    
-
-        <tr>
-            
-                <td class="name"><code>zoomHandler</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type"><a href="module-Zoom-ZoomHandler.html">module:Zoom~ZoomHandler</a></span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The ZoomHandler object.</td>
-        </tr>
-
-    
-    </tbody>
-</table>
-
-
-
-
-
-
-<dl class="details">
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-    <dt class="tag-source">Source:</dt>
-    <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line15">line 15</a>
-    </li></ul></dd>
-    
-
-    
-
-    
-
-    
-</dl>
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
     
     </div>
 
@@ -242,7 +64,7 @@
     
 
     
-    <h4 class="name" id="~init"><span class="type-signature">(inner) </span>init<span class="signature">()</span><span class="type-signature"></span></h4>
+    <h4 class="name" id=".endGroupingSelection"><span class="type-signature">(static) </span>endGroupingSelection<span class="signature">()</span><span class="type-signature"></span></h4>
     
 
     
@@ -250,7 +72,7 @@
 
 
 <div class="description">
-    Initialize handlers and controls and create event listeners.
+    Remove the grouping selection menu.
 </div>
 
 
@@ -294,7 +116,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line32">line 32</a>
+        <a href="Grouping.js.html">Grouping.js</a>, <a href="Grouping.js.html#line16">line 16</a>
     </li></ul></dd>
     
 
@@ -328,7 +150,7 @@
     
 
     
-    <h4 class="name" id="~resetListeners"><span class="type-signature">(inner) </span>resetListeners<span class="signature">()</span><span class="type-signature"></span></h4>
+    <h4 class="name" id=".triggerGroupSelection"><span class="type-signature">(static) </span>triggerGroupSelection<span class="signature">()</span><span class="type-signature"></span></h4>
     
 
     
@@ -336,7 +158,7 @@
 
 
 <div class="description">
-    Reset select event listeners.
+    Trigger the grouping selection menu.
 </div>
 
 
@@ -380,7 +202,93 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line48">line 48</a>
+        <a href="Grouping.js.html">Grouping.js</a>, <a href="Grouping.js.html#line7">line 7</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+        
+            
+
+    
+
+    
+    <h4 class="name" id="~initGroupingListeners"><span class="type-signature">(inner) </span>initGroupingListeners<span class="signature">()</span><span class="type-signature"></span></h4>
+    
+
+    
+
+
+
+<div class="description">
+    The grouping dropdown listener.
+</div>
+
+
+
+
+
+
+
+
+
+
+
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="Grouping.js.html">Grouping.js</a>, <a href="Grouping.js.html#line24">line 24</a>
     </li></ul></dd>
