+++ doc/module-Select.DragSelect.html
@@ -29,7 +29,7 @@
 <header>
     
         <h2><span class="attribs"><span class="type-signature"></span></span>
-            <span class="ancestors"><a href="module-Select.html">Select</a>.</span>DragSelect<span class="signature">(dragHandler, zoomHandler, groupingHandler, neonView)</span><span class="type-signature"></span></h2>
+            <span class="ancestors"><a href="module-Select.html">Select</a>.</span>DragSelect<span class="signature">(dragHandler, zoomHandler, neonView)</span><span class="type-signature"></span></h2>
         
     
 </header>
@@ -42,7 +42,7 @@
     
 
     
-    <h4 class="name" id="DragSelect"><span class="type-signature"></span>new DragSelect<span class="signature">(dragHandler, zoomHandler, groupingHandler, neonView)</span><span class="type-signature"></span></h4>
+    <h4 class="name" id="DragSelect"><span class="type-signature"></span>new DragSelect<span class="signature">(dragHandler, zoomHandler, neonView)</span><span class="type-signature"></span></h4>
     
 
     
@@ -132,29 +132,6 @@
 
         <tr>
             
-                <td class="name"><code>groupingHandler</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type"><a href="GroupingHandler.html">GroupingHandler</a></span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">Instantiated GroupingHandler object.</td>
-        </tr>
-
-    
-
-        <tr>
-            
                 <td class="name"><code>neonView</code></td>
             
 
@@ -212,7 +189,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line222">line 222</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line221">line 221</a>
     </li></ul></dd>
     
 
@@ -390,7 +367,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line375">line 375</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line374">line 374</a>
     </li></ul></dd>
     
 
@@ -476,7 +453,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line268">line 268</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line267">line 267</a>
     </li></ul></dd>
     
 
@@ -562,7 +539,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line289">line 289</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line288">line 288</a>
     </li></ul></dd>
     
 
@@ -648,7 +625,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line240">line 240</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line239">line 239</a>
     </li></ul></dd>
     
 
@@ -852,7 +829,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line394">line 394</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line393">line 393</a>
     </li></ul></dd>
     
 
@@ -896,13 +873,13 @@
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Select.html">Select</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Select.ClickSelect.html">ClickSelect</a></li><li><a href="module-Select.DragSelect.html">DragSelect</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Grouping.html">Grouping</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Select.html">Select</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Select.ClickSelect.html">ClickSelect</a></li><li><a href="module-Select.DragSelect.html">DragSelect</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:46:39 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:54:47 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
